split out model logic into separate modules
