+++ qt/python/mantidqtinterfaces/mantidqtinterfaces/TexturePlanner/presenter.py
@@ -89,7 +89,7 @@ class TexturePlannerPresenter(object):
 
     def set_model_group(self):
         self.model.set_group(self.view.get_group())
-        self.model.get_detQ_lab()
+        self.model.geometry.recompute()
 
     def set_instrument_options(self):
         self.view.set_instrument_options(self.model.get_supported_instruments())
@@ -106,37 +106,37 @@ class TexturePlannerPresenter(object):
         self.view.set_angle_steps()
 
     def set_max_orientation_index(self):
-        num_orientations = self.model.get_num_orientations()
+        num_orientations = self.model.orientations.get_num_orientations()
         self.view.set_max_ind(num_orientations)
 
     def get_vecs(self):
-        return self.model.get_vecs(self.view.get_vecs(), self.view.get_num_gonios())
+        return self.model.orientations.get_vecs(self.view.get_vecs(), self.view.get_num_gonios())
 
     def get_senses(self):
-        return self.model.get_senses(self.view.get_senses(), self.view.get_num_gonios())
+        return self.model.orientations.get_senses(self.view.get_senses(), self.view.get_num_gonios())
 
     def get_angles(self):
-        return self.model.get_angles(self.view.get_angles(), self.view.get_num_gonios())
+        return self.model.orientations.get_angles(self.view.get_angles(), self.view.get_num_gonios())
 
     def on_goniometer_updated(self, goniometer_index, *args):
         self.model.set_gonio_index(goniometer_index)
         orientation_index = self.view.get_current_index()
-        self.model.update_gRs(self.get_vecs(), self.get_senses(), self.get_angles(), orientation_index)
+        self.model.orientations.update_gRs(self.get_vecs(), self.get_senses(), self.get_angles(), orientation_index)
         self.model.update_projected_data(orientation_index)
-        self.model.update_gonio_string(self.get_vecs(), self.get_senses(), self.get_angles(), orientation_index)
+        self.model.orientations.update_gonio_string(self.get_vecs(), self.get_senses(), self.get_angles(), orientation_index)
         self.update_plots()
         self.update_table()
 
     def on_num_gonio_updated(self):
         num_gonios = self.view.get_num_gonios()
-        self.model.set_n_gonio(num_gonios)
+        self.model.orientations.set_n_gonio(num_gonios)
         self.update_enabled_gonios(num_gonios)
         self.on_goniometer_updated(self.model.update_gonio_index(num_gonios))
         self.view.hide_axis_columns()
 
     def on_directions_updated(self):
         self.set_model_texture_directions()
-        self.model.get_detQ_lab()
+        self.model.geometry.recompute()
         self.model.update_all_projected_data()
         self.update_plots()
 
@@ -157,7 +157,7 @@ class TexturePlannerPresenter(object):
         self.update_plots()
 
     def update_plots(self):
-        self.model.update_plot(
+        self.model.plotter.update_plot(
             self.get_vecs(),
             self.get_senses(),
             self.get_angles(),
@@ -168,14 +168,14 @@ class TexturePlannerPresenter(object):
         )
 
     def update_table(self):
-        self.view.tableWidget.setRowCount(self.model.get_num_orientations())
-        for row_index, row_info in enumerate(self.model.get_table_info()):
+        self.view.tableWidget.setRowCount(self.model.orientations.get_num_orientations())
+        for row_index, row_info in enumerate(self.model.orientations.get_table_info()):
             self.view.add_table_row(row_index, *row_info)
 
     def add_orientation(self):
         # add an orientation to the dictionary
-        self.model.add_orientation()
-        new_orientation_index = self.model.get_num_orientations() - 1
+        self.model.orientations.add_orientation()
+        new_orientation_index = self.model.orientations.get_num_orientations() - 1
 
         # update the orientation selector
         self.update_orientation_selector(new_orientation_index)
@@ -183,21 +183,21 @@ class TexturePlannerPresenter(object):
         self.on_goniometer_updated(self.model.gonio_index)
 
     def update_orientation_selector(self, new_index):
-        self.view.spnIndex.setMaximum(self.model.get_num_orientations())
+        self.view.spnIndex.setMaximum(self.model.orientations.get_num_orientations())
         self.view.set_current_index(new_index)
 
     def setup_group_options(self):
         self.view.setup_group_options(self.model.supported_groups)
 
     def update_goniometer_values_from_index(self, index):
-        vecs, senses, angles = self.model.get_goniometer_values(index)
+        vecs, senses, angles = self.model.orientations.get_goniometer_values(index)
         self.view.set_vecs(vecs)
         self.view.set_senses(senses)
         self.view.set_angles(angles)
 
     def on_index_changed(self):
         updated_index = self.view.get_current_index()
-        self.model.set_orientation_index(updated_index)
+        self.model.orientations.set_orientation_index(updated_index)
 
         # we now update the goniometer fields with the saved values
         self.update_goniometer_values_from_index(updated_index)
@@ -215,62 +215,62 @@ class TexturePlannerPresenter(object):
         self.view.set_outputs_enabled(self.view.get_save_dir() != "" and self.view.get_save_filename() != "")
 
     def load_stl(self):
-        self.model.load_stl(self.view.get_stl_string())
+        self.model.workspaces.load_stl(self.view.get_stl_string())
         self.set_initial_shape()
 
     def load_xml(self):
-        self.model.load_xml(self.view.get_xml_string())
+        self.model.workspaces.load_xml(self.view.get_xml_string())
         self.set_initial_shape()
 
     def load_orientation_file(self):
-        num_gonios = self.model.load_orientation_file(self.view.get_orientation_file())
+        num_gonios = self.model.orientations.load_orientation_file(self.view.get_orientation_file())
         # update selected number of goniometers to match
         self.view.set_num_gonios(num_gonios)
         # update the orientation selector
-        self.view.set_current_index(self.model.get_num_orientations() - 1)
-        self.update_orientation_selector(self.model.get_num_orientations())
+        self.view.set_current_index(self.model.orientations.get_num_orientations() - 1)
+        self.update_orientation_selector(self.model.orientations.get_num_orientations())
         # update table and plot
         self.model.update_all_projected_data()
         self.update_table()
         self.update_plots()
 
     def update_selected(self):
-        self.model.update_selected(self.view.get_select_indices())
+        self.model.orientations.update_selected(self.view.get_select_indices())
 
     def update_included(self):
-        self.model.update_included(self.view.get_include_indices())
+        self.model.orientations.update_included(self.view.get_include_indices())
         self.update_plots()
 
     def select_all(self):
-        self.model.select_all()
+        self.model.orientations.select_all()
         self.update_table()
 
     def deselect_all(self):
-        self.model.deselect_all()
+        self.model.orientations.deselect_all()
         self.update_table()
 
     def delete_selected(self):
-        self.model.delete_selected()
-        self.view.set_current_index(self.model.get_orientation_index())
+        self.model.orientations.delete_selected()
+        self.view.set_current_index(self.model.orientations.get_orientation_index())
         # call on index change, just in case the index happens to remain, we still need the same update
         self.on_index_changed()
-        self.update_orientation_selector(self.model.get_orientation_index())
+        self.update_orientation_selector(self.model.orientations.get_orientation_index())
         # update table and plot
         self.model.update_all_projected_data()
         self.update_table()
         self.update_plots()
 
     def to_sscanss(self):
-        self.model.output_as_sscanss(self.view.get_save_dir(), self.view.get_save_filename())
+        self.model.exporter.output_as_sscanss(self.view.get_save_dir(), self.view.get_save_filename())
 
     def to_euler(self):
-        self.model.output_as_euler(self.view.get_save_dir(), self.view.get_save_filename())
+        self.model.exporter.output_as_euler(self.view.get_save_dir(), self.view.get_save_filename())
 
     def to_matrix(self):
-        self.model.output_as_matrix(self.view.get_save_dir(), self.view.get_save_filename())
+        self.model.exporter.output_as_matrix(self.view.get_save_dir(), self.view.get_save_filename())
 
     def to_reference_workspace(self):
-        self.model.output_as_reference_workspace(self.view.get_save_dir(), self.view.get_save_filename())
+        self.model.exporter.output_as_reference_workspace(self.view.get_save_dir(), self.view.get_save_filename())
 
     def set_show_mu(self):
         self.update_custom_shape_finder_enabled()
@@ -280,7 +280,7 @@ class TexturePlannerPresenter(object):
         self.update_plots()
 
     def set_initial_shape(self):
-        self.model.update_initial_shape(
+        self.model.workspaces.update_initial_shape(
             self.view.get_init_x(),
             self.view.get_init_y(),
             self.view.get_init_z(),
@@ -292,12 +292,12 @@ class TexturePlannerPresenter(object):
         self.update_plots()
 
     def set_gauge_volume(self):
-        self.model.set_gauge_volume_str(self.view.get_shape_method(), self.view.get_custom_shape())
+        self.model.workspaces.set_gauge_volume_str(self.view.get_shape_method(), self.view.get_custom_shape())
         self.model.update_all_projected_data()
         self.update_plots()
 
     def clear_gauge_volume(self):
-        self.model.set_gauge_volume_str("No Gauge Volume", None)
+        self.model.workspaces.set_gauge_volume_str("No Gauge Volume", None)
         self.model.update_all_projected_data()
         self.update_plots()
